requirements.txt
praw~=7.7.1
python-dotenv~=1.0.1
psycopg2-binary
git+https://github.com/alabeybahri/tagme-python.git
requests~=2.31.0
nltk~=3.8.1
spacy~=3.7.4
